Changed network module interface. See below:
Instead of having three channels:

* outgoing     - Send to a specified machine ID
* outgoing_all - Send to all machines in subnet
* incoming     - Get any incoming message

we now have five channels:

* to_master      - Send to specified machine tagged FROM_CLIENT
* to_any_master  - Broadcast to all machines tagged FROM CLIENT
* to_all_clients - Broadcast to all machines tagged FROM MASTER
* from_master    - Receive message tagged FROM MASTER
* from_client    - Receive message tagged FROM CLIENT

The reason for this, is because the machine running the master
routine must also run a client routine. For example, consider
the client who is designated backup. When it detects a master
loss event, it will need to become the master. It would be weird
to stop being a client in that case, since any orders that it
was doing would need to be aborted.

Anyway, a machine must be able to be master and client concurrently.
However, when there was only one INCOMING channel, we had no way
to seperate messages that the master was supposed to read, from the
messages that the client was supposed to read. That is, both do

select {
    case <- incoming:
        ...
}

To fix this, I introduced tagging of network packets (a single byte
that is appended to any outgoing packets, and parsed out from any
incoming packets).