Broadcast works at the lab. But there are issues.
When we broadcast, we also receive an echo on our own machine.
This is picked up by the network module, which proceeds to
send a message over the FROM_CLIENT channel. But when we are
running an instance of only the Client routine, we do not read
from this channel. Thus, the network module becomes blocked
from sending further messages.

We can not just read from the channel in the Client routine,
since it will run concurrently with the Master routine - which
DOES read from the channel.

In main I added a concurrent thread that reads from the channel
when we only launch the client. This fixed the bugs I had today,
but will not fix the bugs we get in the future when a backup
client needs to become the master. Since then we need to stop
that routine, and it gets really ugly.

TODO: Figure this out.