Clients don't need to send their targetfloor.
Upon writing this I realize that clients might not need
to send the TargetFloor in the ClientData. We can deduce
their current target by inspecting the list of orders. If
the list has no priority order, then the lift is idle.

In fact, this might be a better approach. Consider what
happens when a lift goes black when taking an order. When
it resets it should be able to pick up where it left off.

That might cause a bug in the current implementation.
(Upon startup targetfloor will be set to 0. If the client
sends a packet containing this value, before synchronizing
with the master list... uh oh!).