Fixed bug in DistributeWork function.
The bug was that the lift might not actually change its already
prioritized floor. Which is a perfectly normal scenario, but I
forgot to consider it.

Also fixed bug relating to having multiple requests, and the
deletion of requests. The bug was the traditional delete element
from list while iterating over it and forget to decrement the
iteration index.

Also added ability to clear all button lamps. Which we do when
we get an update from the master.