Fixed a bug related to backup takeover.
When a backup client took over as master, it did not reassign the
orders that the master's client had been assigned. Made a temp fix
for this by reassigning upon takeover.

Ideally, we should reassign orders whenever they have been inactive
for too long. That would handle this case anyway.