Master now syncs Client list as well.
That is, the MasterData struct is now
	struct MasterData
	{
		AssignedBackup network.ID
		Orders []Order
		Clients map[network.ID]Client
	}

The motivation for this is that when the backup client takes over,
it has no information regarding the clients that have been
assigned to the different orders. They may have timed out since
the last time the master checked!

Indeed, when the master dies, its client dies with it. So there's
one case already! But because the backup didn't have a TimeoutTimer
associated with the already existing clients, the clients who die
are left hanging.

This fixes that, by making the backup keep track of the client
list as well. When the master starts up, it checks if it has
any backup information, and sync with the client list if any.
It also launches the ListenForClientTimeout threads.

Now wait a minute... why can't you just launch the ListenFor...
threads given the IDs in the orders list - that we already have?

The reason behind that is that I would have to add those as
clients in the client map. But I do not know the lifts' LPF, so
the DistributeWork function could potentially screw up.