Fixed bug on client.
The target floor must be set to INVALID_FLOOR at the beginning.
It was set to 0, which screwed up things if we got an order for
first floor at startup.

Furthermore, I had forgotten to actually set the clients map
equal to that which the client gets from the master.