Added logging system.
Orders are now considered complete when the lift opens the door,
as opposed to when it closes the door. Not sure if this is a good
idea, but it solves one timing issue:

When an order is done, it is added to the request pile. But since
it takes time before the master sees that the order was done, it
will send us old data saying that the order is not done. The lift
would then do the order multiple times.